requirements.txt
tensorflow_hub
tensorflow
openai
numpy==1.25.2
requests
bs4
playwright
datetime
tiktoken
PyPDF2
scidownl
futures==3.1.1
mysql-connector-python
flask
